Move RoutingTripPattern to transit model
