+++ src/test/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/request/TripPatternWithId.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.routing.algorithm.raptoradapter.transit.request;
 
-import org.opentripplanner.routing.algorithm.raptoradapter.transit.RoutingTripPattern;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
+import org.opentripplanner.transit.model.network.RoutingTripPattern;
 import org.opentripplanner.transit.model.network.TripPattern;
 
 public class TripPatternWithId extends RoutingTripPattern {
